perf(reactivity): optimize graph allocations with scratch pools and refine owner lifecycle
Motivation & Context:
Reactive graph propagation, node disposal, and dependency tracking
previously allocated multiple transient vectors, sets, and slotmap edges
on every run and subscription change. Under high-frequency churn and deep
graph traversal, these allocations introduced substantial overhead and
memory fragmentation. Additionally, dynamically staged dependencies modified
during an active computation could miss subsequent notifications if written
before transaction finalization, and the owner registry needed robust
generation recycling with a wider u64 counter.

Key Technical Details:
- Introduce reusable `PropagationScratch`, `DisposalScratch`, `DependencyBuffer`,
  and `TargetBuffer` scratch pools to eliminate allocations in hot paths.
- Simplify `NodeAdjacency` by storing direct `HashSet<TargetNode>` instances
  and remove dedicated `ReactiveEdge` / `EdgeId` slotmaps.
- Implement `stage_pending_source` and `propagate_pending_sources` in eval and
  graph runtime to handle write-after-read replay for newly observed dependencies.
- Refactor scheduler owner tracking into `OwnerSlotTable` with explicit slot
  states (`Vacant`, `Active`, `Inactive`, `Retired`), safe idempotent release,
  and upgrade generation counters from `u32` to `u64`.
- Wrap scratch executions with `catch_unwind` and `resume_unwind` to guarantee
  scratch reset and pool return even in the event of panics.
- Refine reactivity benchmarks to categorize propagation frontier topologies
  and separate slot churn from full owner churn.

Side Effects & Risks:
- Owner ID generations are now 64-bit (`u64`), updating internal registry
  signatures and runtime snapshots.
- Scratch pool buffers are retained up to a defined limit (`BUFFER_RETAIN_LIMIT`)
  to balance amortized reuse against peak memory retention.

Verification:
- Added graph tests covering read-then-write new dependency replays, rollback on
  callback error, and ordering semantics (`tests/graph.rs`).
- Added unit tests for bidirectional adjacency uniqueness, scratch pool reuse,
  high-water tracking, and `OwnerSlotTable` generation retirement (`runtime/graph.rs`,
  `runtime/scheduler.rs`).
- Updated Criterion benchmark suites covering propagation frontiers and owner churn.

+++ crates/silex_reactivity/src/unsafe_boundary.rs
@@ -156,7 +156,7 @@ impl<'scope> ActiveOwnerProof<'scope> {
     /// restored only after all registry and state checks have succeeded.
     pub(crate) fn from_registry(
         id: OwnerId,
-        generation: u32,
+        generation: u64,
         expected: &WeakOwnerToken,
         state: Rc<ErasedScopeState>,
     ) -> ReactiveResult<Option<Self>> {
@@ -353,7 +353,7 @@ pub(crate) struct CleanupOwnerProof<'scope> {
 impl<'scope> CleanupOwnerProof<'scope> {
     pub(crate) fn from_registry(
         id: OwnerId,
-        generation: u32,
+        generation: u64,
         expected: &WeakOwnerToken,
         state: Rc<ErasedScopeState>,
     ) -> ReactiveResult<Option<Self>> {
